Remove Level unit test
This is an integration test, not a unit test.  This test does not test any part of the unit, it only tests if level 1-1 exists and Level does not throw an error.  A level needs to be synthetically fed to the level for a proper unit test, or better yet, no level needs to be fed into level for the level to make a level.